+++ olcrtc/script/docker/olcrtc-healthcheck.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-set -eu
-
-pidof olcrtc >/dev/null 2>&1